% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdx_basic.r
\name{cdx_basic_query}
\alias{cdx_basic_query}
\title{Perform a basic/limited Internet Archive CDX resource query for a URL}
\usage{
cdx_basic_query(url)
}
\arguments{
\item{url}{URL to query}
}
\description{
Perform a basic/limited Internet Archive CDX resource query for a URL
}
\examples{
cdx_basic_query("rud.is/b")
}

