% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memento.r
\name{get_mementos}
\alias{get_mementos}
\title{Retrieve site mementos}
\usage{
get_mementos(url, timestamp)
}
\arguments{
\item{url}{URL to retrieve information for}

\item{timestamp}{(optional) timestamp use when checking for availability}
}
\description{
Retrieve site mementos
}
