% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ia-fields.R
\name{ia_fields}
\alias{ia_fields}
\title{Return a character vector of available return fields for \code{\link[=ia_scrape]{ia_scrape()}}}
\usage{
ia_fields()
}
\description{
Return a character vector of available return fields for \code{\link[=ia_scrape]{ia_scrape()}}
}
