% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ia-retrieve.R
\name{ia_retrieve}
\alias{ia_retrieve}
\title{Retrieve directory listings for Internet Archive objects by identifier}
\usage{
ia_retrieve(identifier)
}
\arguments{
\item{identifier}{an identifier string obtained via \code{\link[=ia_scrape]{ia_scrape()}}.}
}
\value{
data frame of files and links with some metadata
}
\description{
Given an object identifier (obtained via \code{\link[=ia_scrape]{ia_scrape()}}), retrieve the
directory listing for the item. The \code{link} column will have URLs that
can be retrieved with \code{download.file()}.
}
\examples{
nasa <- ia_scrape("collection:nasa", count=100L)
item <- ia_retrieve(nasa$identifier[1])
td <- tempdir()
download.file(item$link[1], file.path(td, item$file[1]))
}
