% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{is_memento}
\alias{is_memento}
\alias{is_first_memento}
\alias{is_next_memento}
\alias{is_prev_memento}
\alias{is_last_memento}
\alias{is_original}
\alias{is_timemap}
\alias{is_timegate}
\title{Various helper tests for memento \code{rel} values}
\usage{
is_memento(x)

is_first_memento(x)

is_next_memento(x)

is_prev_memento(x)

is_last_memento(x)

is_original(x)

is_timemap(x)

is_timegate(x)
}
\arguments{
\item{x}{character vector (ideally from a \code{rel} \code{memento} \code{data_frame} column)}
}
\value{
\code{logical} vector
}
\description{
Very useful in a \code{purrr} or \code{data_frame} context.
}
