% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ia-scrape.R
\name{print.ia_scrape}
\alias{print.ia_scrape}
\title{Print method for \code{ia_scrape} objects}
\usage{
\method{print}{ia_scrape}(x, ...)
}
\arguments{
\item{x}{an \code{ia_scrape} object}

\item{...}{unused}
}
\description{
Print method for \code{ia_scrape} objects
}
\keyword{internal}
