% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mementoweb.r
\name{read_memento}
\alias{read_memento}
\title{Read a resource directly from the Time Travel MementoWeb}
\usage{
read_memento(url, timestamp = format(Sys.Date(), "\%Y"), as = c("text",
  "raw", "parsed"))
}
\arguments{
\item{url}{URL to retrieve information for}

\item{timestamp}{timestamp to use when checking for availability.If you don't pass in a
valid R "time-y" object, you will need to ensure the character string you
provide is in a valid subset of \code{YYYYMMDDhhmmss}.}

\item{as}{How you want the content returned. One of "\code{text}", "\code{raw}" or "\code{parsed}"
(it uses \code{httr::content()} to do the heavy lifting).}
}
\value{
The specified content type
}
\description{
This uses the MementoWeb \href{http://timetravel.mementoweb.org/about/}{Time Travel Service}
to directly retrieve retrieve URL content. You get (slightly augmented by MementoWeb)
orignal URL content from at or near the timestamp you specify.
}
\details{
This can be helpful if you find the resource you need is gone today but still
exists in the memory of the internet.
}
\examples{
\dontrun{
library(htmltools)
library(rvest)

yahoo_old <- read_memento("http://yahoo.com", "2010")

html_print(HTML(yahoo_old)) # opens browser

pg <- xml2::read_html(yahoo_old)
html_nodes(pg, "a.x3-large") \%>\%
  html_text()
## Tiger loses one more sponsorship"
}
}
\references{
\href{http://timetravel.mementoweb.org/guide/api/}{Time Travel API}
}
