% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slideshow.r
\name{slideshow}
\alias{slideshow}
\title{Generate a htmlwidgets slideshow of memento snapshots}
\usage{
slideshow(url)
}
\arguments{
\item{url}{URL to retrieve information for}
}
\value{
slick.js image carousel htmlwidget
}
\description{
\strong{WARNING}: this function requires that you have correctly set up:
\itemize{
\item docker
\item splashr
}
}
\examples{
\dontrun{
slideshow("http://rstats.org")
}
}
