% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\docType{package}
\name{wayback}
\alias{wayback}
\alias{wayback-package}
\title{Tools to Work with the Various Internet Archive Wayback Machine APIs}
\description{
The 'Internet Archive' provides access to millions of cached sites. Methods
are provided to access these cached resources through the 'APIs' provided by
the 'Internet Archive'.
}
\author{
Bob Rudis (bob@rud.is)
}
