% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wayback.R
\name{wayback_scrape}
\alias{wayback_scrape}
\title{Scrape wayback url's using the rvest package}
\usage{
wayback_scrape(url, css, xpath)
}
\arguments{
\item{url}{Wayback Machine url}

\item{css}{Nodes to select.}

\item{xpath}{Nodes to select.}
}
\description{
Scrape wayback url's using the rvest package
}
\examples{
wayback_scrape(wayback_url(url =
 "https://www.python.org/psf/members/",
 timestamp = "20140101")$url,
 "#board-of-directors li , #board-of-directors h1")
}

