% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wayback.R
\name{wayback_url}
\alias{wayback_url}
\title{Query wayback machine API}
\usage{
wayback_url(url, timestamp = NULL)
}
\arguments{
\item{url}{character url}

\item{timestamp}{character date of 1-14 digits in YYYYMMDDhhmmss format}
}
\description{
Query wayback machine API
}
\examples{
wayback_url(url = "example.com", timestamp = "20140101")
}

