% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_install.R
\name{wb_install}
\alias{wb_install}
\title{wb_pkg_install}
\usage{
wb_install(pkg, repo_url, library_path, package_path, dependencies)
}
\arguments{
\item{pkg}{\code{character} - package name specified as cha}

\item{repo_url}{url of repository to get packages from}

\item{library_path}{\code{character} - lib_path to use first}

\item{package_path}{\code{character} - path to look for binary packages to install}

\item{dependencies}{\verb{character vector} - specifies which types of dependencies
should be installed also}
}
\description{
Installs packages and its dependencies
}
