% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_pkg_base_packages.R
\name{wb_pkg_base_packages}
\alias{wb_pkg_base_packages}
\title{wb_pkg_base_packages}
\usage{
wb_pkg_base_packages()
}
\value{
character vector with packages already part of the R installation
}
\description{
wb_pkg_base_packages
}
