% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_pkg_dependencies.R
\name{wb_pkg_dependencies}
\alias{wb_pkg_dependencies}
\title{wb_pkg_dependencies}
\usage{
wb_pkg_dependencies(
  pkg,
  library_path = NULL,
  dep_types = c("imports", "depends", "linkingto")
)
}
\arguments{
\item{dep_types}{}
}
\description{
wb_pkg_dependencies
}
