% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_pkg_install.R
\name{wb_pkg_install}
\alias{wb_pkg_install}
\title{wb_pkg_install}
\usage{
wb_pkg_install(
  pkg,
  repo_url,
  library_path,
  package_path,
  dependencies,
  tasks = NULL
)
}
\arguments{
\item{dependencies}{}
}
\description{
wb_pkg_install
}
