% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_pkg_install_worker.R
\name{wb_pkg_install_worker}
\alias{wb_pkg_install_worker}
\title{wb_pkg_install_worker}
\usage{
wb_pkg_install_worker(pkg, repo_url, package_path, library_path, dependencies)
}
\arguments{
\item{pkg}{pkg to install}

\item{repo_url}{repo url to install from}

\item{dependencies}{}
}
\value{

}
\description{
wb_pkg_install_worker
}
\examples{

}
