% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_require.R
\name{wb_require}
\alias{wb_require}
\title{wb_require}
\usage{
wb_require(
  pkg,
  date = options("wb_require_date"),
  date_shift = options("wb_require_shift_date"),
  lib_path = options("wb_require_lib_path"),
  binary_path = options("wb_require_binary_path"),
  dependencies = getOption("wb_require_dependencies", c("Depends", "Imports",
    "LinkingTo"))
)
}
\arguments{
\item{pkg}{\code{character} - package name specified as cha}

\item{date}{\code{character} - date specified in ISO format, defaults to date
specified in \code{sessionInfo()$R.version$version.string}}

\item{date_shift}{\code{integer} - number of days to add to date}

\item{lib_path}{\code{character} - lib_path to use first}

\item{binary_path}{\code{character} - path to look for binary packages to install}
}
\description{
This function will make sure a specified package is available within the
current R session. If the package is not available it will attempt to
install it and load it thereafter. If its already installed it will only
load the package.
}
\details{
Other than the normal \code{install.packages} function this function will not
use the newest package available but one that was available at a specific
point in time on CRAN. For CRAN snapshots the Microsoft R Open snapshots and
the checkpoiunt package are used.

The default date is the date of the release of the R version currently in use
plus an additional 60 days.

The \code{lib_path} argument can be used to make use a project specific
R package library path.

Also, its possible to specify a path were binary packages can be found and
installed from.
}
\examples{

\dontrun{

# make sure apckage is available
wb_require("data.table")

# use packages released up to 365 days younger than the current R-versions
# release date
wb_require("data.table", date_shift = 365)

# use a specific lib path for loading and installing
wb_require("data.table", lib_path = "./r_lib_path")

}

}
