% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.r
\name{archive_available}
\alias{archive_available}
\title{Does the Internet Archive have a URL cached?}
\usage{
archive_available(url, timestamp)
}
\arguments{
\item{url}{URL to retrieve information for}

\item{timestamp}{(optional) timestamp use when checking for availability}
}
\description{
Does the Internet Archive have a URL cached?
}

