% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdx_basic.r
\name{cdx_basic_query}
\alias{cdx_basic_query}
\title{Perform a basic/limited Internet Archive CDX resource query for a URL}
\usage{
cdx_basic_query(url, match_type = c("exact", "prefix", "host", "domain"),
  collapse = "urlkey", limit = 10000L)
}
\arguments{
\item{url}{URL/resource to query for}

\item{match_type}{The CDX server can also return results matching a certain
prefix, a certain host or all subdomains. Can be one of
\code{"exact"}, \code{"prefix"}, \code{"host"}, or \code{"domain"} (defaults to \code{exact}).}

\item{collapse}{collapse results based on a field, or a substring of a field.
Collapsing is done on adjacent cdx lines where all captures after the
first one that are duplicate are filtered out. This is useful for filtering
out captures that are 'too dense' or when looking for unique captures.
To use collapsing, add one or more \code{collapse=field} or \code{collapse=field:N}
where \code{N} is the first \code{N} characters of field to test. Use \code{NULL} for
no collapsing Default is to collapse by \code{urlkey} (like the web UX).
Reference: \url{https://github.com/internetarchive/wayback/tree/master/wayback-cdx-server}.}

\item{limit}{Maximum number of results to return (first \emph{n} results). Use a
negative number to retrieve the last \emph{n} results. Default is \code{10,000}.}
}
\value{
data frame
}
\description{
CDX files are "Content Index" files. The Wayback CDX server is a standalone HTTP
servlet that serves the index that the Wayback machine uses to lookup captures.
}
\details{
The index format is known as 'cdx' and contains various fields representing the capture,
usually sorted by url and date. \url{http://archive.org/web/researcher/cdx_file_format.php}.
}
\examples{
\dontrun{
rproj_basic <- cdx_basic_query("https://www.r-project.org/")

dplyr::glimpse(rproj_basic)
## Observations: 10,000
## Variables: 7
## $ urlkey     <chr> "org,r-project)/", "org,r-project)/", "org,r-project)/"...
## $ timestamp  <dttm> 2000-06-20, 2000-08-16, 2000-10-12, 2000-11-10, 2000-1...
## $ original   <chr> "http://www.r-project.org:80/", "http://www.r-project.o...
## $ mimetype   <chr> "text/html", "text/html", "text/html", "text/html", "te...
## $ statuscode <chr> "200", "200", "200", "200", "200", "200", "200", "200",...
## $ digest     <chr> "XDIHHFDLIWSZFHYHT453ZL5FYPCKFF6Z", "SRO3WSKQS6HST4PQY7...
## $ length     <dbl> 4894, 5027, 589, 581, 582, 596, 590, 592, 592, 592, 563...
}
}
