% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memento.r
\name{get_mementos}
\alias{get_mementos}
\title{Retrieve site mementos}
\usage{
get_mementos(url, timestamp)
}
\arguments{
\item{url}{URL to retrieve information for}

\item{timestamp}{(optional) timestamp use when checking for availability}
}
\description{
It's possible that a memento list could be large and contain an entry
for the next "page" of mementos.
}
\details{
TODO Not sure we should handle that ^^ here since this is a free service and
is already pretty taxed. If the user needs it they can request from a new
timestamp. But I'm not committed to this concept.
}
