% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timemap.r
\name{get_timemap}
\alias{get_timemap}
\title{Retrieve a timemap for a URL}
\usage{
get_timemap(url)
}
\arguments{
\item{url}{Either an existig timemap URL or a plain resource URL}
}
\description{
Retrieve a timemap for a URL
}
