% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_pkg_install_worker.R
\name{wb_pkg_install_worker}
\alias{wb_pkg_install_worker}
\title{wb_pkg_install_worker}
\usage{
wb_pkg_install_worker(pkg, repo_url, package_path, library_path, dependencies)
}
\arguments{
\item{pkg}{\code{character} - package name specified as cha}

\item{repo_url}{url of repository to get packages from}

\item{package_path}{\code{character} - path to look for binary packages to install}

\item{library_path}{\code{character} - lib_path to use first}

\item{dependencies}{\verb{character vector} - specifies which types of dependencies
should be installed also}
}
\description{
wb_pkg_install_worker
}
